Create La3a
The “Fizz Buzz Problem” is a well-known interview question. It usually is given as something like the following:
Write a program that prints the numbers from 1 to 100. But for multiples of three print
“Fizz” instead of the number and for the multiples of five print “Buzz”. For numbers that
are multiples of both three and five print “FizzBuzz”.
In this assignment, you’ll be implementing a variation of this problem to make sure you are comfortable
writing a simple Java program from scratch using command line arguments